properly TAA ssr

+++ gle/voxelisation.cpp
@@ -9,13 +9,13 @@ voxel::grid voxel::create_grid(glm::ivec3 resolution, aabb bb)
 	grid.voxel_unit = glm::vec3((aabb_dim.x / resolution.x), (aabb_dim.y / resolution.y), (aabb_dim.z / resolution.z));
 	grid.resolution = resolution;
 	grid.bounding_box = bb;	
-	grid.voxel_texture = texture::create_3d_texture_empty(resolution, GL_RGBA, GL_RGBA32F, GL_FLOAT);
-	glAssert(glBindImageTexture(0, grid.voxel_texture.m_handle, 0, GL_TRUE, 0, GL_READ_WRITE, GL_RGBA32F));
-	glAssert(glBindImageTexture(1, grid.voxel_texture.m_handle, 1, GL_TRUE, 0, GL_READ_WRITE, GL_RGBA32F));
-	glAssert(glBindImageTexture(2, grid.voxel_texture.m_handle, 2, GL_TRUE, 0, GL_READ_WRITE, GL_RGBA32F));
-	glAssert(glBindImageTexture(3, grid.voxel_texture.m_handle, 3, GL_TRUE, 0, GL_READ_WRITE, GL_RGBA32F));
-	glAssert(glBindImageTexture(4, grid.voxel_texture.m_handle, 4, GL_TRUE, 0, GL_READ_WRITE, GL_RGBA32F));
-	glAssert(glBindImageTexture(5, grid.voxel_texture.m_handle, 5, GL_TRUE, 0, GL_READ_WRITE, GL_RGBA32F));
+	grid.voxel_texture = texture::create_3d_texture_empty(resolution, GL_RGBA, GL_RGBA16F, GL_FLOAT);
+	glAssert(glBindImageTexture(0, grid.voxel_texture.m_handle, 0, GL_TRUE, 0, GL_READ_WRITE, GL_RGBA16F));
+	glAssert(glBindImageTexture(1, grid.voxel_texture.m_handle, 1, GL_TRUE, 0, GL_READ_WRITE, GL_RGBA16F));
+	glAssert(glBindImageTexture(2, grid.voxel_texture.m_handle, 2, GL_TRUE, 0, GL_READ_WRITE, GL_RGBA16F));
+	glAssert(glBindImageTexture(3, grid.voxel_texture.m_handle, 3, GL_TRUE, 0, GL_READ_WRITE, GL_RGBA16F));
+	glAssert(glBindImageTexture(4, grid.voxel_texture.m_handle, 4, GL_TRUE, 0, GL_READ_WRITE, GL_RGBA16F));
+	glAssert(glBindImageTexture(5, grid.voxel_texture.m_handle, 5, GL_TRUE, 0, GL_READ_WRITE, GL_RGBA16F));
 
 	return grid;
 }
